+++ WinterEngine.DataAccess/Repositories/ItemRepository.cs
@@ -28,42 +28,43 @@ namespace WinterEngine.DataAccess
 
         #region Methods
 
-        /// <summary>
-        /// Adds an item to the database.
-        /// </summary>
-        /// <param name="item">The item to add to the database.</param>
-        /// <returns></returns>
-        public Item Add(Item item)
+        private Item InternalSave(Item item, bool saveChanges)
         {
-            return _context.Items.Add(item);
-        }
-
-        public void Add(List<Item> itemList)
-        {
-            _context.Items.AddRange(itemList);
-        }
-
-        /// <summary>
-        /// If an item with the same resref is in the database, it will be replaced with newItem.
-        /// If an item does not exist by newItem's resref, it will be added to the database.
-        /// </summary>
-        /// <param name="newItem">The new item to upsert.</param>
-        public Item Save(Item item)
-        {
+            Item retItem;
             if (item.ResourceID <= 0)
             {
-                _context.Items.Add(item);
+                retItem = _context.Items.Add(item);
             }
             else
             {
-                Update(item);
+                retItem = _context.Items.SingleOrDefault(x => x.ResourceID == item.ResourceID);
+                if (retItem == null) return null;
+                _context.Entry(retItem).CurrentValues.SetValues(item);
+
             }
-            return item;
+            if (saveChanges)
+            {
+                _context.SaveChanges();
+            }
+
+            return retItem;
+        }
+
+        public Item Save(Item item)
+        {
+            return InternalSave(item, true);
         }
 
         public void Save(IEnumerable<Item> entityList)
         {
-            throw new NotImplementedException();
+            if (entityList != null)
+            {
+                foreach (var item in entityList)
+                {
+                    InternalSave(item, false);
+                }
+                _context.SaveChanges();
+            }
         }
 
         /// <summary>
@@ -92,43 +93,42 @@ namespace WinterEngine.DataAccess
             _context.LocalVariables.RemoveRange(dbItem.LocalVariables.ToList());
             _context.LocalVariables.AddRange(newItem.LocalVariables.ToList());
             
-        }        
+        }
 
-        /// <summary>
-        /// Deletes and item from the database.
-        /// </summary>
-        /// <param name="item"></param>
+        private void DeleteInternal(Item item, bool saveChanges = true)
+        {
+            var dbItem = _context.Items.SingleOrDefault(x => x.ResourceID == item.ResourceID);
+            if (dbItem == null) return;
+
+            _context.LocalVariables.RemoveRange(item.LocalVariables.ToList());
+
+            _context.Items.Remove(dbItem);
+
+            if (saveChanges)
+            {
+                _context.SaveChanges();
+            }
+        }
+
         public void Delete(Item item)
         {
-            this.Delete(item.ResourceID);
+            DeleteInternal(item);
         }
 
-        /// <summary>
-        /// Deletes an item with the specified resref from the database.
-        /// </summary>
-        /// <param name="resref">The resource reference to search for and delete.</param>
-        /// <returns></returns>
         public void Delete(int resourceID)
         {
-            Item item = _context.Items.Where(i => i.ResourceID == resourceID).SingleOrDefault();
-            _context.LocalVariables.RemoveRange(item.LocalVariables.ToList());
-            _context.Items.Remove(item);
+            var item = _context.Items.Find(resourceID);
+            DeleteInternal(item);
         }
 
-        public void Delete(IEnumerable<Item> entityList)
+        public void Delete(IEnumerable<Item> itemList)
         {
-            throw new NotImplementedException();
+            foreach (var item in itemList)
+            {
+                DeleteInternal(item, false);
+            }
+            _context.SaveChanges();
         }
-
-        /// <summary>
-        /// Returns all of the items from the database.
-        /// </summary>
-        /// <returns></returns>
-        public IEnumerable<Item> GetAll()
-        {
-            return _context.Items.ToList();
-        }
-        
         //todo: move this logic somewhere else
         //public List<DropDownListUIObject> GetAllUIObjects()
         //{
@@ -143,15 +143,16 @@ namespace WinterEngine.DataAccess
         //    return items;
         //}
 
+        public IEnumerable<Item> GetAll()
+        {
+            return _context.Items.AsEnumerable();
+        }
+
         public Item GetByID(int resourceID)
         {
             return _context.Items.Where(x => x.ResourceID == resourceID).SingleOrDefault();
         }
 
-        public void ApplyChanges()
-        {
-            _context.SaveChanges();
-        }
 
         /// <summary>
         /// Returns all of the items in a specified category from the database.
